Create StickyView component for MyObs
