+++ src/components/MyObservations/MyObservations.js
@@ -1,16 +1,13 @@
 // @flow
 import Header from "components/MyObservations/Header";
-import { ObservationsFlashList, ViewWrapper } from "components/SharedComponents";
+import { ObservationsFlashList, StickyView, ViewWrapper } from "components/SharedComponents";
 import { View } from "components/styledComponents";
 import type { Node } from "react";
 import React, { useRef, useState } from "react";
 import { Animated, Platform } from "react-native";
-import { useDeviceOrientation } from "sharedHooks";
 
 import LoginSheet from "./LoginSheet";
 import MyObservationsEmpty from "./MyObservationsEmpty";
-
-const { diffClamp } = Animated;
 
 type Props = {
   isFetchingNextPage?: boolean,
@@ -35,11 +32,6 @@ const MyObservations = ( {
   showLoginSheet,
   setShowLoginSheet
 }: Props ): Node => {
-  const {
-    isTablet,
-    screenHeight,
-    screenWidth
-  } = useDeviceOrientation( );
   const [heightAboveToolbar, setHeightAboveToolbar] = useState( 0 );
 
   const [hideHeaderCard, setHideHeaderCard] = useState( false );
@@ -47,22 +39,6 @@ const MyObservations = ( {
   // basing collapsible sticky header code off the example in this article
   // https://medium.com/swlh/making-a-collapsible-sticky-header-animations-with-react-native-6ad7763875c3
   const scrollY = useRef( new Animated.Value( 0 ) );
-
-  // On Android, the scroll view offset is a double (not an integer), and interpolation shouldn't be
-  // one-to-one, which causes a jittery header while slow scrolling (see issue #634).
-  // See here as well: https://stackoverflow.com/a/60898411/1233767
-  const scrollYClamped = diffClamp(
-    scrollY.current,
-    0,
-    heightAboveToolbar * 2
-  );
-
-  // Same as comment above (see here: https://stackoverflow.com/a/60898411/1233767)
-  const offsetForHeader = scrollYClamped.interpolate( {
-    inputRange: [0, heightAboveToolbar * 2],
-    // $FlowIgnore
-    outputRange: [0, -heightAboveToolbar]
-  } );
 
   const renderEmptyList = ( ) => <MyObservationsEmpty isFetchingNextPage={isFetchingNextPage} />;
 
@@ -94,16 +70,7 @@ const MyObservations = ( {
     <>
       <ViewWrapper>
         <View className="overflow-hidden">
-          <Animated.View
-            style={[
-              {
-                transform: [{ translateY: offsetForHeader }],
-                height: isTablet
-                  ? screenHeight
-                  : Math.max( screenWidth, screenHeight )
-              }
-            ]}
-          >
+          <StickyView scrollY={scrollY} heightAboveView={heightAboveToolbar}>
             <Header
               toggleLayout={toggleLayout}
               layout={layout}
@@ -125,7 +92,7 @@ const MyObservations = ( {
               data={observations.filter( o => o.isValid() )}
               showObservationsEmptyScreen
             />
-          </Animated.View>
+          </StickyView>
         </View>
       </ViewWrapper>
       {showLoginSheet && <LoginSheet setShowLoginSheet={setShowLoginSheet} />}
